v1.0: Implemented AI prediction to main program, fixed pen not responding during prediction, cleaned project structures.
