--- functions/__init__.py
@@ -1,5 +1,0 @@
-from .parsing import *
-from .symbol_detection import *
-
-
-__all__ = ["get_symbol_bounding", "pixels_isolation", "separate_instructions", "perform_calc", "perform_priority_calc"]
