+++ scripts/preprocessing.py
@@ -2,7 +2,7 @@ import tensorflow as tf
 import matplotlib.pyplot as plt
 import numpy as np
 from tensorflow.keras.preprocessing.image import img_to_array
-from configuration import batch_size, img_height, img_width
+from configuration import BATCH_SIZE, IMG_HEIGHT, IMG_WIDTH
 
 
 data_augmentation = tf.keras.Sequential([
@@ -25,6 +25,7 @@ def preprocess_images(image, label):
     image = tf.cast(image, tf.float32) / 255.0
     return image, label
 
+
 def preprocess_dataset(dataset, batch_size, augment=False):
     if augment:
         dataset = dataset.map(lambda x, y: (data_augmentation(x), y), num_parallel_calls=tf.data.AUTOTUNE)
@@ -35,15 +36,19 @@ def preprocess_dataset(dataset, batch_size, augment=False):
 
 
 def preprocess_image_main(image, img_height, img_width):
-    if isinstance(image, np.ndarray):  # Si c'est un tableau NumPy
+    # Si c'est un tableau NumPy
+    if isinstance(image, np.ndarray): 
         # Si l'image a trois canaux (couleur), convertir en niveaux de gris
         if image.shape[-1] == 3:
             image = tf.image.rgb_to_grayscale(image)
-        image = tf.image.resize(image, (img_height, img_width))
-    else:  # Sinon, on suppose que c'est une image PIL
+
+        
+    # Sinon, on suppose que c'est une image PIL
+    else:  
         image = image.resize((img_height, img_width))
         image = img_to_array(image) # Convertir en tableau NumPy
         image = tf.image.rgb_to_grayscale(image)
+
     # Normalisation et ajout de dimension batch
     image = tf.cast(image, tf.float32) / 255.0
     image = tf.expand_dims(image, axis=0)
