+++ scripts/test_data_loader.py
@@ -1,15 +1,15 @@
 import os
 import tensorflow as tf
 from scripts.data_loader import load_data
-from configuration import img_height, img_width, data_dir, project_root
+from configuration import IMG_HEIGHT, IMG_WIDTH, DATA_DIR, PROJECT_ROOT
 
-print("chemin data_dir : ", data_dir)
-data_bad = os.path.join(project_root, "handwritting_math_calculation", "images")
+print("chemin data_dir : ", DATA_DIR)
+data_bad = os.path.join(PROJECT_ROOT, "handwritting_math_calculation", "images")
 
 
 def test_load_data():
     try:
-        dataset, class_names = load_data(data_dir, img_height, img_width, batch_size=None)
+        dataset, class_names = load_data(DATA_DIR, IMG_HEIGHT, IMG_WIDTH, batch_size=None)
         assert isinstance(dataset, tf.data.Dataset), "Le dataset n'est pas du bon type"
         print("Le dataset est du bon type")
         assert isinstance(class_names, list), "Les noms de classes ne sont pas dans une liste"
@@ -22,7 +22,7 @@ def test_load_data():
 
         total_images = tf.data.experimental.cardinality(dataset).numpy()
         print(f"Total d'images : {total_images}")
-        class_counts = {class_name: len(os.listdir(os.path.join(data_dir, class_name))) for class_name in os.listdir(data_dir)}
+        class_counts = {class_name: len(os.listdir(os.path.join(DATA_DIR, class_name))) for class_name in os.listdir(DATA_DIR)}
         print(class_counts)
         
         # Vérification des classes
@@ -35,14 +35,14 @@ def test_load_data():
 
     # Test 2 : Chemin invalide
     try:
-        load_data("path/that/does/not/exist", img_height, img_width)
+        load_data("path/that/does/not/exist", IMG_HEIGHT, IMG_WIDTH)
         print("Test 2 échoué : Une exception était attendue.")
     except Exception:
         print("Test 2 réussi.")
 
     # Test 3 : Structure incorrecte
     try:
-        dataset, class_names = load_data(data_bad, img_height, img_width)
+        dataset, class_names = load_data(data_bad, IMG_HEIGHT, IMG_WIDTH)
         assert len(class_names) > 0, "Aucune classe trouvée"
         print("Test 3 réussi.")
     except Exception as e:
